# user_service/openapi-spec.yaml
openapi: "3.0.3"
info:
  title: "User Service API"
  description: "Service for registering, logging in and editing user data"
  version: "1.0"
paths:
  /register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "Successfully registered"
        "400":
          description: "Registration failed"
  /login:
    post:
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "Login succeded"
          headers:
            Set-Cookie:
              description: "Access token"
              schema:
                type: string
        "400":
          description: "Login failed"
  /update:
    post:
      summary: Update user info
      parameters:
        - name: "Cookie"
          in: "Header"
          description: "Cookie with auth token"
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  nullable: true
                last_name:
                  type: string
                  nullable: true
                date_of_birth:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
                phone_number:
                  type: string
                  nullable: true
      responses:
        "200":
          description: "Update succeded"
        "400":
          description: "Update failed"